refatorado relacionamento entre as entidades

+++ backend/src/main/java/com/barberbross/BarberBross/dto/request/DTOServicoRequest.java
@@ -3,6 +3,8 @@ package com.barberbross.BarberBross.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record DTOServicoRequest(
         @NotBlank
         String nome,
@@ -12,7 +14,7 @@ public record DTOServicoRequest(
         Long categoriaId,
 
         @NotNull
-        double preco,
+        BigDecimal preco,
 
         @NotNull
         int duracao) {}
